+++ README.md
@@ -1,4 +1,4 @@
-![spirus](/public/made_with_spirus.png?height=100)
+![spirus](/spirus_logo.png?height=100)
 # Spirus
 HTML5 Electron Typescript Canvas framework thing, scrapped off one of my games called Yushite.
