--- public/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Spirus</title>
-	<script src="dist/bundle.js" defer></script>
-</head>
-<body>
-	<canvas id="canvas"></canvas>
-	<div id="watermark">
-		<a href="https://github.com/SpirusDev/Spirus" target="_blank">
-			<img src="./made_with_spirus.png" alt="Made with Spirus" style="position: absolute; bottom: 0; left: 0;" width="135" height="85">
-		</a>
-	</div>
-</body>
-</html>
