+++ spirus.config.js
@@ -6,7 +6,7 @@ module.exports = {
 		height: 1080
 	},
 	discordRPC: {
-		enabled: true, // feel free to disable
+		enabled: false, // feel free to disable
 		clientID: "", // set to your client ID
 
 		config: {
@@ -14,12 +14,9 @@ module.exports = {
 			details: "https://github.com/EclipseLikesSpace/Spirus", // the details
 			startTimestamp: Date.now(),
 			endTimestamp: null, // leave null for "Time elapsed"
-			largeImageKey: 'spirus_large', // need to set these in your Discord Developer Portal
-			smallImageKey: 'spirus_small', // ^
+			largeImageKey: "spirus_large", // need to set these in your Discord Developer Portal
+			smallImageKey: "spirus_small", // ^
 			instance: true, // leave this true
 		}
 	},
-	spirusOptions: {
-		madeWithSpirusWatermark: true // disable this if it is blocking canvas
-	}
 }
