--- src/entry.ts
@@ -1,5 +1,0 @@
-// Typescript
-import "./ts/index.ts";
-
-// Stylesheets
-import "./scss/style.scss"
