+++ tsconfig.json
@@ -1,15 +1,25 @@
 {
 	"compilerOptions": {
-		"outDir": "./public/dist/",
-		"noImplicitAny": true,
-		"module": "es6",
-		"target": "es5",
-		"allowJs": true,
-		"resolveJsonModule": true,
-		"moduleResolution": "node",
-		"lib": [
-			"es2018",
-			"DOM"
-		]
-	}
+        "outDir": "./dist",
+		"pretty": true,
+        "lib": ["es2020", "DOM"],
+        "target": "es2017",
+        "moduleResolution": "node",
+
+        "declaration": true,
+        "declarationMap": true,
+        "sourceMap": true,
+        "composite": true,
+        "noEmitOnError": true,
+
+        "strict": true,
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
+
+        "skipLibCheck": true,
+
+        "alwaysStrict": true,
+        "preserveConstEnums": true,
+        "newLine": "lf",
+    }
 }
